Cleaned up error generation and added token cache
- Both Authorization and Access token endpoints now use the same
error builder. Also renamed types to have more appropriate names

- Added an authorization token store to the sample application. The
store is used to store the generated token, for a maximum, of 10
minutes (following the recommendations in the specification). The
store is used when generating an authorization token and later to
verify that the token is valid before returning an access token to
the caller.